Move some code from retroarch.c over to menu_driver.c
